+++ trapeze/src/test/java/com/jkjamies/trapeze/TrapezeTest.kt
@@ -39,7 +39,7 @@ private data class ScreenB(val id: Int = 0) : TrapezeScreen {
 
 private class FakeStateHolder : TrapezeStateHolder<TrapezeScreen, TrapezeState, TrapezeEvent>() {
     @Composable
-    override fun produceState(screen: TrapezeScreen): TrapezeState {
+    override fun produceState(): TrapezeState {
         return object : TrapezeState {}
     }
 }
@@ -49,10 +49,9 @@ private val FakeUi: TrapezeUi<TrapezeState> = @Composable { _: Modifier, _: Trap
 class TrapezeTest : BehaviorSpec({
 
     Given("a Trapeze built with factories for ScreenA") {
-        val stateHolder = FakeStateHolder()
         val trapeze = Trapeze.Builder()
             .addStateHolderFactory { screen, _ ->
-                if (screen is ScreenA) stateHolder else null
+                if (screen is ScreenA) FakeStateHolder() else null
             }
             .addUiFactory { screen ->
                 if (screen is ScreenA) FakeUi else null